WIP Template project file auto-creator
We have a template.html file that we want to read from as an example for which our other Auto-created HTML pages pull from only editing specific HTML elements for specific bits of information.

Therefore, the end result would be auto-generated HTML pages that are identical in CSS style and format but different in what they comprise of. Name of the files will also need to be changed.